requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0
whitenoise>=6.6.0
openai>=1.3.0
google-generativeai>=0.3.0
Pillow>=10.0.0
PyPDF2>=3.0.0
# Celery and Redis are NOT supported on Vercel serverless — excluded
# celery>=5.3.0
# redis>=5.0.0
